cd into directories before uploading/downloading
Re-work scripts to `cd` into appropriate working directories and then
invoke `mirror` without passing the `source` or `target` args. I.e.,
just mirror the current directory on the client to the current directory
on the server (or vice-versa, in the case of `/download`).

The most important thing this addresses is, from what I can tell, a bug
in lftp where it attempts to `mkdir` the `$REMOTE_PATH` on upload, but
doesn't gracefully handle the (admittedly slightly funky) server telling
it it cannot do that. It *also* addresses some funky semantics that
`lftp` has around paths that end with `/` when mirroring (the presence /
nonpresence of a trailing `/` will make `lftp` think it needs to modify
the path in some way). I feel like we're explicitly building a blunt instrument
here--"make that directory over there be like this directory over
here"--so I'd rather have it not mangle my paths in ways that I cannot
reason about without returning to the documentation every single time.

A side effect of this change is that for `/upload` the
target directory *must already exist on the server* and for `/download`
the target directory *must already exist on the client*. Again, blunt
instrument.